added formatter.tsx and useDriveAttach.ts hook
