+++ lib/formatters.tsx
@@ -1,0 +1,25 @@
+  export const formatDate = (dateString: string) => {
+    const date = new Date(dateString);
+    const userTimezoneOffset = date.getTimezoneOffset() * 60000;
+    const adjustedDate = new Date(date.getTime() + userTimezoneOffset);
+    return adjustedDate.toLocaleDateString("en-US", {
+      month: "short",
+      day: "numeric",
+      year: "numeric",
+    });
+  };
+
+  export const formatAmount = (amount: number) => {
+    const isIncome = amount > 0;
+    const formatted = new Intl.NumberFormat("en-US", {
+      style: "currency",
+      currency: "USD",
+      minimumFractionDigits: 2,
+    }).format(Math.abs(amount));
+    return (
+      <span>
+        {isIncome ? "+" : "-"}
+        {formatted}
+      </span>
+    );
+  };
